Separate utils into files & io (network requests)
