+++ src/arguments.rs
@@ -13,9 +13,9 @@ use rand::{Rng, distributions::Alphanumeric};
 use gethostname::gethostname;
 use url::form_urlencoded;
 use webbrowser;
-use crate::{settings::Settings, utils};
+use crate::{settings::Settings, utils, files, io};
 
-const AUTH_TOKEN_LENGHT: u8 = 69;
+const AUTH_TOKEN_LENGTH: u8 = 69;
 const SCAN_FILE_TYPE: [&str; 7] = ["sql", "rs", "py", "js", "css", "html", "svelte"];
 const DOWNLOAD_EXCLUDE: [&str; 3] = ["exam", "project", "graduation"];
 
@@ -77,7 +77,7 @@ fn open_ide(current_attempt: &Attempt, editors: &Vec<String>) -> () {
             Err(_) => {},
         }
     }
-
+    
     for editor in &editors {
         let mut editor_parts = editor.split_whitespace();
         let editor_name = editor_parts.next().unwrap_or_default();
@@ -141,10 +141,10 @@ fn install_logic() {
 fn grade_logic(settings: &Settings, arg: String) {
     let token = settings.config.get("auth", "token").unwrap_or("".to_string());
     let url_arg = format!("/api/attempts/{}", arg.replace("~", ":"));
-    let response = utils::request("GET", url_arg, &token, None);
+    let response = io::request("GET", url_arg, &token, None);
 
     let attempts = match response {
-        Some(data) => utils::response_to_json(data),
+        Some(data) => io::response_to_json(data),
         None => {
             eprintln!("No attempt found");
             exit(1)
@@ -153,10 +153,10 @@ fn grade_logic(settings: &Settings, arg: String) {
 
     let attempt = &attempts[0];
 
-    let out_dir = utils::get_lms_dir().join("grading").join(attempt.get("spec").unwrap().as_str().unwrap().to_string().replace(":", "~"));
+    let out_dir = files::get_lms_dir().join("grading").join(attempt.get("spec").unwrap().as_str().unwrap().to_string().replace(":", "~"));
 
     if Path::exists(&out_dir) {
-        if utils::is_folder_empty(&out_dir).unwrap() {
+        if files::is_folder_empty(&out_dir).unwrap() {
             match fs::remove_dir_all(&out_dir) {
                 Ok(_) => {},
                 Err(err) => eprintln!("Cant remove directory because: {}", err)
@@ -169,7 +169,7 @@ fn grade_logic(settings: &Settings, arg: String) {
     } else {
         let _ = fs::create_dir_all(&out_dir);
         let url = format!("/api/attempts/{}/submission", attempt.get("spec").unwrap().as_str().unwrap().to_string());
-        utils::download_tgz(url, &token, &out_dir);
+        io::download_tgz(url, &token, &out_dir);
         println!("Downloaded to {}", out_dir.to_str().unwrap().to_string());
     }
 
@@ -213,7 +213,7 @@ fn grade_logic(settings: &Settings, arg: String) {
 fn login_logic(mut settings: Settings) {
     let token: String = rand::thread_rng()
         .sample_iter(&Alphanumeric)
-        .take(AUTH_TOKEN_LENGHT.into())
+        .take(AUTH_TOKEN_LENGTH.into())
         .map(char::from)
         .collect();
 
@@ -264,7 +264,7 @@ fn upload_logic(settings: &Settings) {
 
 
 
-    if utils::is_folder_empty(&current_attempt.path).unwrap() {
+    if files::is_folder_empty(&current_attempt.path).unwrap() {
         if !utils::prompt_yes_no("This folder is currently empty are you sure you want to upload?") {
             exit(0)
         }
@@ -293,9 +293,9 @@ fn upload_logic(settings: &Settings) {
 
     let url = format!("/api/attempts/{}/submission", current_attempt.id.to_string());
 
-    match utils::request("POST", url, &token, Some(data.stdout)) {
+    match io::request("POST", url, &token, Some(data.stdout)) {
         Some(res) => {
-            let json_res: serde_json::Value = utils::response_to_json(res);
+            let json_res: serde_json::Value = io::response_to_json(res);
 
             match json_res.get("transferred") {
                 Some(transferred) => {
@@ -326,9 +326,9 @@ fn download_logic(settings: &Settings, arg: String) {
         let _  = download_attempt(&arg, &token);
     }
 
-    let response = utils::request("GET", "/api/node-paths".to_string(), &token, None);
+    let response = io::request("GET", "/api/node-paths".to_string(), &token, None);
     let attempts = match response {
-        Some(data) => utils::response_to_json(data),
+        Some(data) => io::response_to_json(data),
         None => {
             eprintln!("No attempt found");
             exit(1)
@@ -336,7 +336,7 @@ fn download_logic(settings: &Settings, arg: String) {
     };
 
     let mut local_dirs: HashSet<String> = HashSet::new();
-    let target_dir = utils::get_lms_dir().join("*/*");
+    let target_dir = files::get_lms_dir().join("*/*");
     for path in  glob(target_dir.to_str().unwrap()).expect("Faild to read lms dir") {
         match path {
             Ok(path) => {
@@ -366,10 +366,10 @@ fn download_logic(settings: &Settings, arg: String) {
 
 fn download_attempt(assignment: &String, token: &String) -> bool {
     let url_arg = format!("/api/attempts/@{}", assignment.replace("~", ":"));
-    let response = utils::request("GET", url_arg, token, None);
+    let response = io::request("GET", url_arg, token, None);
 
     let attempts = match response {
-        Some(data) => utils::response_to_json(data),
+        Some(data) => io::response_to_json(data),
         None => {
             eprintln!("No attempt found: {}", assignment);
             return false
@@ -380,7 +380,7 @@ fn download_attempt(assignment: &String, token: &String) -> bool {
 
     match attempt.as_object() {
         Some(select_attempt) => {
-            let mut out_dir = utils::get_lms_dir();
+            let mut out_dir = files::get_lms_dir();
 
             match select_attempt.get("path") {
                 Some(att) => {
@@ -397,7 +397,7 @@ fn download_attempt(assignment: &String, token: &String) -> bool {
                     let _ = fs::create_dir_all(&out_dir);
 
                     let url = format!("/api/attempts/{}/submission", select_attempts.as_str().unwrap());
-                    utils::download_tgz(url, &token, &out_dir);
+                    io::download_tgz(url, &token, &out_dir);
                     println!("Downloaded: {} at: {}", assignment, &out_dir.to_str().unwrap());
                 }
                 None => return false 
@@ -426,12 +426,12 @@ fn template_logic(settings: &Settings) {
 
 
 fn get_current_attempt(token: String) -> Attempt {
-    let mut lms_dir = utils::get_lms_dir();
+    let mut lms_dir = files::get_lms_dir();
 
     let mut cache = lms_dir.clone();
     cache.push(".cache");
 
-    let res = utils::request("GET", "/api/attempts/current".to_string(), &token, None);
+    let res = io::request("GET", "/api/attempts/current".to_string(), &token, None);
 
     if res.is_none() {
         if Path::exists(&cache) {
@@ -452,7 +452,7 @@ fn get_current_attempt(token: String) -> Attempt {
         exit(1)
     }
 
-    let online_attempt = utils::response_to_json(res.unwrap());
+    let online_attempt = io::response_to_json(res.unwrap());
     let assignment_path = &online_attempt;
 
     if assignment_path.is_null() {
@@ -483,7 +483,7 @@ fn download_template(token: String, attempt: &Attempt) -> bool {
         let _ = fs::create_dir_all(&attempt.path);
         println!("Created {}", &attempt.path.to_str().unwrap());
     } else {
-        if !utils::is_folder_empty(&attempt.path).unwrap() {
+        if !files::is_folder_empty(&attempt.path).unwrap() {
             return false
         }
     }
@@ -494,7 +494,7 @@ fn download_template(token: String, attempt: &Attempt) -> bool {
     }
 
     let url = format!("/api/attempts/{}/template", &attempt.id);
-    utils::download_tgz(url, &token, &attempt.path);
+    io::download_tgz(url, &token, &attempt.path);
     true
 }
 
@@ -505,10 +505,10 @@ fn verify_logic() {
 }
 
 fn move_node_directories() -> bool {
-    let lms_dir = utils::get_lms_dir();
+    let lms_dir = files::get_lms_dir();
 
-    let correct_pathes_json = match utils::request("GET", "/api/node-paths".to_string(), &"".to_string(), None) {
-        Some(data) => utils::response_to_json(data),
+    let correct_paths_json = match io::request("GET", "/api/node-paths".to_string(), &"".to_string(), None) {
+        Some(data) => io::response_to_json(data),
         None => {
             eprintln!("Cant convert paths to json");
             exit(1)
@@ -533,8 +533,8 @@ fn move_node_directories() -> bool {
                 }
 
                 // TODO: Refactor this
-                if let correct_path_object = Some(&correct_pathes_json) {
-                    let pressent_node_id = correct_pathes_json.as_object().unwrap().get(&node_id);
+                if let correct_path_object = Some(&correct_paths_json) {
+                    let pressent_node_id = correct_paths_json.as_object().unwrap().get(&node_id);
                     if pressent_node_id.is_some() {
                         if let correct_path = pressent_node_id.unwrap().as_str().unwrap().to_string() {
